+++ changelog.md
@@ -1,3 +1,9 @@
-# v1.0.0 - release Interactions update 
+# v1.0.1 - release Interactions update 
 
-added missing commands
+- Fixed a bug where if /play did not answer in time it would crash
+- Cleaned the code a bit
+- Created a cool webpage
+- added time print option in config.py
+- added command for module disabling
+- added command for feedback
+-